refactor(container/profile): refactor code to use common component
use a function to conditionaly render UserSetUp components
