+++ App.js
@@ -9,7 +9,7 @@ import SignInScreen from "./screens/SignInScreen";
 import MailSignInScreen from "./screens/MailSignInScreen";
 import OTPScreen from "./screens/OTPScreen";
 import SignUpScreen from "./screens/SignUpScreen";
-import Profile from "./containers/profile/Profile";
+import UserSetup from "./containers/profile/UserSetup";
 const Stack = createNativeStackNavigator();
 
 export default function App() {
@@ -21,7 +21,7 @@ export default function App() {
             <Stack.Screen name="Home" component={HomeScreen} />
             <Stack.Screen name="SignUp" component={SignUpScreen} />
             <Stack.Screen name="SignIn" component={SignInScreen} />
-            <Stack.Screen name="MailSignIn" component={MailSignInScreen} />
+            <Stack.Screen name="MailSignIn" component={UserSetup} />
             <Stack.Screen name="OTP" component={OTPScreen} />
           </Stack.Group>
         </Stack.Navigator>
